Inline fixture renderer
